Moved app.js and app.spec.js into /tests #5
